Breaking changes: - because of huge numerical errors related to small values of probabilities we had to drop our internal usage of decimal type and switch back to double type -Suffix and Prefix algorithms done -Some other bugfixes

+++ HMM-DishonestCasino/Casino/LoadedDice.cs
@@ -7,7 +7,7 @@ namespace HMMDishonestCasino.Casino
 {
     public class LoadedDice : Dice
     {
-        private const decimal MAX_UNFAIR_PROBABILITY = 0.5m;
+        private const double MAX_UNFAIR_PROBABILITY = 0.5;
 
         public LoadedDice(int numberOfSides, Random random) : base(numberOfSides, random)
         {
@@ -43,7 +43,7 @@ namespace HMMDishonestCasino.Casino
 
         public override int Roll()
         {
-            var randomValue = (decimal) random.NextDouble();
+            var randomValue = (double) random.NextDouble();
 
             var distribution = GetDistribution();
